fix: image generating in demos

+++ demo/app.py
@@ -9,9 +9,9 @@ import numpy as np
 # Device and dtype configuration
 def get_device_and_dtype():
     if torch.cuda.is_available():
-        return 'cuda', torch.bfloat16
+        return 'cuda', torch.float32
     elif torch.backends.mps.is_available():
-        return 'mps', torch.float16
+        return 'mps', torch.float32
     return 'cpu', torch.float32
 
 device, dtype = get_device_and_dtype()
